renamed api route to match zone model
